requirements.txt
pandas>=1.3.0
openpyxl>=3.0.0
anthropic>=0.5.0
openai>=1.0.0
google-generativeai>=0.3.0
pytest>=7.0.0
requests>=2.28.0